+++ tests/one_taskset.rs
@@ -18,6 +18,8 @@ const ABOUT_KEY: usize = 1;
 
 // Load test configuration.
 const EXPECT_WORKERS: usize = 2;
+const USERS: usize = 3;
+const RUN_TIME: usize = 3;
 
 // There are multiple test variations in this file.
 #[derive(Clone)]
@@ -121,12 +123,6 @@ fn validate_one_taskset(
         TestType::ResetMetrics => {
             // Statistics were reset after all users were started, so Goose should report
             // fewer page loads than the server actually saw.
-            println!(
-                "response_time_counter: {}, mock_endpoint_called: {}",
-                index_metrics.response_time_counter,
-                mock_endpoints[INDEX_KEY].times_called()
-            );
-
             assert!(index_metrics.response_time_counter < mock_endpoints[INDEX_KEY].times_called());
             assert!(
                 index_metrics.status_code_counts[&status_code]
@@ -228,12 +224,22 @@ fn run_gaggle_test(test_type: TestType) {
                 "--manager",
                 "--expect-workers",
                 &EXPECT_WORKERS.to_string(),
+                "--users",
+                &USERS.to_string(),
                 "--no-reset-metrics",
             ],
         ),
         TestType::ResetMetrics => common_build_configuration(
             &server,
-            &mut vec!["--manager", "--expect-workers", &EXPECT_WORKERS.to_string()],
+            &mut vec![
+                "--manager",
+                "--expect-workers",
+                &EXPECT_WORKERS.to_string(),
+                "--users",
+                &USERS.to_string(),
+                "--run-time",
+                &RUN_TIME.to_string(),
+            ],
         ),
     };
 
@@ -273,8 +279,6 @@ fn test_one_taskset_reset_metrics() {
     run_standalone_test(TestType::ResetMetrics);
 }
 
-/* @TODO: @FIXME: Goose is not resetting metrics when running in Gaggle mode.
- * Issue: https://github.com/tag1consulting/goose/issues/193
 #[test]
 #[cfg_attr(not(feature = "gaggle"), ignore)]
 #[serial]
@@ -283,4 +287,3 @@ fn test_one_taskset_reset_metrics() {
 fn test_one_taskset_reset_metrics_gaggle() {
     run_gaggle_test(TestType::ResetMetrics);
 }
-*/