+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - add `test/sequence.rs` to confirm sequencing tests works correctly, even in Gaggle mode
  - deduplicate test logic by moving shared functionality into `tests/common.rs`; consistently test functionality both in standalone and Gaggle mode
  - properly create debug log when enabled in Gaggle mode
+ - properly reset metrics after all users launch in Gaggle mode (unless --no-reset-metrics is enabled)
 
 ## 0.10.2 Sep 27, 2020
  - remove unnecessary `GooseAttack.number_of_cpus` instead calling `num_cpus::get()` directly